+++ tests/fields/test_time.py
@@ -44,9 +44,11 @@ class TestDatetimeFields(test.TestCase):
         self.assertEqual(obj2.datetime_add, obj.datetime_add)
 
     async def test_update(self):
-        obj0 = await testmodels.DatetimeFields.create(datetime=datetime(2019, 9, 1, 0, 0, 0))
+        obj0 = await testmodels.DatetimeFields.create(
+            datetime=datetime(2019, 9, 1, 0, 0, 0, tzinfo=get_default_timezone())
+        )
         await testmodels.DatetimeFields.filter(id=obj0.id).update(
-            datetime=datetime(2019, 9, 1, 6, 0, 8)
+            datetime=datetime(2019, 9, 1, 6, 0, 8, tzinfo=get_default_timezone())
         )
         obj = await testmodels.DatetimeFields.get(id=obj0.id)
         self.assertEqual(obj.datetime, datetime(2019, 9, 1, 6, 0, 8, tzinfo=get_default_timezone()))