+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Fixed
 ^^^^^
 - Fix `bulk_create` error. (#1012)
 - Fix unittest invalid.
+- Fix `bulk_update` in `postgres` with some type like `datetime`. (#968)
 
 0.18.0
 ------