+++ CHANGELOG.rst
@@ -12,9 +12,9 @@ Changelog
 ------
 Fixed
 ^^^^^
-- Fix `bulk_create` error. (#1012) (#1022)
+- Fix `bulk_create` error. (#1012)
 - Fix unittest invalid.
-- Fix `bulk_update` in `postgres` with some type like `datetime`. (#968)
+- Fix `bulk_update` in `postgres` with some type. (#968) (#1022)
 
 0.18.0
 ------